Add files via upload
Project  I did as a hobby in 11th grade. It is a simple processing script that reads serial data from a USB port in the computer or the raspberry pi serial port, and displays it in a GUI that makes it easy to read for non programmers.  Addition of storing received data on a text document was later added. This project was specifically designed to to display 3 values; temperature, air humidity and light percentage. Both temperature and humidity were read by the same Arduino sensor(dht22), the light percentage was read by a photocell.